ForcePluginsTests is now SalesforceHybridTest and lives under libs/test So only the lib SalesforceReact is without tests
