+++ .travis.yml
@@ -26,6 +26,8 @@ script:
    - ./gradlew :libs:SalesforceSDK:connectedAndroidTest
    - ./gradlew :libs:SmartStore:connectedAndroidTest
    - ./gradlew :libs:SmartSync:connectedAndroidTest
+   - ./gradlew :libs:SalesforceHybrid:assembleDebug
+   - ./gradlew :libs:SalesforceReact:assembleDebug
+   - ./gradlew :libs:test:SaleforceHybridTest:connectedAndroidTest
    - ./gradlew :native:SampleApps:RestExplorer:connectedAndroidTest
    - ./gradlew :native:TemplateApp:connectedAndroidTest
-   - ./gradlew :hybrid:test:ForcePluginsTest:connectedAndroidTest
