+++ libs/test/SalesforceHybridTest/AndroidManifest.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="utf-8"?>
 
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-	package="com.salesforce.androidsdk.forcepluginstests"
+	package="com.salesforce.androidsdk.phonegap"
     android:versionCode="1"
     android:versionName="1.0">
 
@@ -15,7 +15,7 @@
         android:targetSdkVersion="23" />
 
     <application android:label="@string/app_name"
-        android:name="com.salesforce.androidsdk.phonegap.app.ForcePluginsTestApp"
+        android:name="com.salesforce.androidsdk.phonegap.app.SalesforceHybridTestApp"
 		android:icon="@drawable/sf__icon">
 
 		<uses-library android:name="android.test.runner" />        
@@ -23,7 +23,7 @@
         <!--  Main activity -->
         <activity
             android:label="@string/app_name"
-            android:name="com.salesforce.androidsdk.phonegap.ui.ForcePluginsTestActivity"
+            android:name="com.salesforce.androidsdk.phonegap.ui.SalesforceHybridTestActivity"
             android:configChanges="orientation|keyboardHidden">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
