+++ libs/test/SalesforceHybridTest/assets/www/cordova.js
