+++ libs/test/SalesforceHybridTest/assets/www/test/SFAbstractSmartStoreTestSuite.js
