+++ libs/test/SalesforceHybridTest/assets/www/test/SFSDKInfoTestSuite.js
