+++ libs/test/SalesforceHybridTest/assets/www/test/SFSmartStoreTestSuite.js
