+++ libs/test/SalesforceHybridTest/assets/www/test/SFSmartSyncTestSuite.js
