+++ libs/test/SalesforceHybridTest/res/drawable-hdpi/icon.png
