+++ libs/test/SalesforceHybridTest/res/drawable-ldpi/icon.png
