+++ libs/test/SalesforceHybridTest/res/drawable-mdpi/icon.png
