+++ libs/test/SalesforceHybridTest/res/values/strings.xml
