+++ libs/test/SalesforceHybridTest/res/xml/config.xml
