+++ settings.gradle
@@ -4,8 +4,8 @@ include 'libs:SmartStore'
 include 'libs:SmartSync'
 include 'libs:SalesforceHybrid'
 include 'libs:SalesforceReact'
+include 'libs:test:SalesforceHybridTest'
 include 'native:TemplateApp'
-include 'reactnative:ReactNativeTemplateApp'
 include 'native:NativeSampleApps:RestExplorer'
 include 'native:NativeSampleApps:SmartSyncExplorer'
 include 'native:NativeSampleApps:AppConfigurator'
@@ -13,4 +13,4 @@ include 'native:NativeSampleApps:ConfiguredApp'
 include 'hybrid:HybridSampleApps:AccountEditor'
 include 'hybrid:HybridSampleApps:NoteSync'
 include 'hybrid:HybridSampleApps:SmartSyncExplorerHybrid'
-include 'hybrid:test:ForcePluginsTest'
+include 'reactnative:ReactNativeTemplateApp'
