+++ tools/sdk.sh
@@ -59,7 +59,7 @@ usage ()
     echo "        SmartSyncExplorer"
     echo "        TemplateApp"
     echo "        ReactNativeTemplateApp"
-    echo "        ForcePluginsTest"
+    echo "        SalesforceHybridTest"
     echo "    <test_target> can be "
     echo "        all"
     echo "        RestExplorerTest"
@@ -67,7 +67,7 @@ usage ()
     echo "        SmartStoreTest"
     echo "        SmartSyncTest"
     echo "        TemplateAppTest"
-    echo "        ForcePluginsTest"
+    echo "        SalesforceHybridTest"
 }
 
 verbose ()
@@ -169,7 +169,6 @@ else
         build_project_if_requested    "SalesforceHybrid"              :libs:SalesforceHybrid
         build_project_if_requested    "SalesforceReact"               :libs:SalesforceReact
         build_project_if_requested    "TemplateApp"                   :native:TemplateApp
-        build_project_if_requested    "ReactNativeTemplateApp"        :reactnative:ReactNativeTemplateApp
         build_project_if_requested    "RestExplorer"                  :native:NativeSampleApps:RestExplorer 
         build_project_if_requested    "AppConfigurator"               :native:NativeSampleApps:AppConfigurator
         build_project_if_requested    "ConfiguredApp"                 :native:NativeSampleApps:ConfiguredApp
@@ -177,7 +176,8 @@ else
         build_project_if_requested    "AccountEditor"                 :hybrid:HybridSampleApps:AccountEditor
         build_project_if_requested    "NoteSync"                      :hybrid:HybridSampleApps:NoteSync
         build_project_if_requested    "SmartSyncExplorerHybrid"       :hybrid:HybridSampleApps:SmartSyncExplorerHybrid
-        build_project_if_requested    "ForcePluginsTest"              :hybrid:test:ForcePluginsTest
+        build_project_if_requested    "SalesforceHybridTest"          :libs:test:SalesforceHybridTest
+        build_project_if_requested    "ReactNativeTemplateApp"        :reactnative:ReactNativeTemplateApp
     fi
 
     if ( should_do "test{all}" )
@@ -185,11 +185,11 @@ else
         header "Testing all"
         ./gradlew connectedAndroidTest  | grep "$TEST_OUTPUT_FILTER"
     else
-        run_test_project_if_requested "SalesforceSDKTest"   :libs:SalesforceSDK
-        run_test_project_if_requested "SmartStoreTest"      :libs:SmartStore
-        run_test_project_if_requested "SmartSyncTest"       :libs:SmartSync
-        run_test_project_if_requested "TemplateAppTest"     :native:TemplateApp
-        run_test_project_if_requested "RestExplorerTest"    :native:NativeSampleApps:RestExplorer
-        run_test_project_if_requested "ForcePluginsTest"    :hybrid:test:ForcePluginsTest
+        run_test_project_if_requested "SalesforceSDKTest"    :libs:SalesforceSDK
+        run_test_project_if_requested "SmartStoreTest"       :libs:SmartStore
+        run_test_project_if_requested "SmartSyncTest"        :libs:SmartSync
+        run_test_project_if_requested "SalesforceHybridTest" :libs:test:SalesforceHybridTest
+        run_test_project_if_requested "TemplateAppTest"      :native:TemplateApp
+        run_test_project_if_requested "RestExplorerTest"     :native:NativeSampleApps:RestExplorer
     fi
 fi
